Plan 9 from Bell Labs 1999-05-07

+++ port/devsd.c
@@ -42,7 +42,7 @@ struct Disk
 {
 	QLock;
 	Target*	t;
-	uchar	lun;
+	int	lun;
 	char	id[NAMELEN];
 	char	vol[NAMELEN];
